refactor(datadog-rum): moves service init to app-init service and checks if datadog has been initialized before adding or removing a user session
