+++ src/app/app.component.ts
@@ -1,7 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { Title } from '@angular/platform-browser';
 import { Router } from '@angular/router';
-import { DatadogRumService } from './main/services/datadog-rum.service';
 
 @Component({
     selector: 'app-root',
@@ -15,15 +14,11 @@ export class AppComponent implements OnInit {
 
     constructor(
         private _router: Router,
-        private _titleService: Title,
-        private _datadogRumService: DatadogRumService
+        private _titleService: Title
     ) {
 
         // set the page title
         this._titleService.setTitle('DaSCH Service Platform');
-
-        // init datadog RUM
-        this._datadogRumService.initializeRum();
     }
 
     ngOnInit() {
